+++ src/app.module.ts
@@ -11,9 +11,9 @@ import { AlbumController } from './album/album.controller';
 import { AlbumService } from './album/album.service';
 import { FavoritesService } from './favorites/favorites.service';
 import { FavoritesController } from './favorites/favorites.controller';
-import { LoggerMiddleware } from './logger.middleware';
-import { LoggingService } from './logging.service';
-import { HttpExceptionFilter } from './http-exception.filter';
+import { LoggerMiddleware } from './logging/logger.middleware';
+import { LoggingService } from './logging/logging.service';
+import { HttpExceptionFilter } from './logging/http-exception.filter';
 import { APP_FILTER } from '@nestjs/core';
 import { JwtModule } from '@nestjs/jwt';
 import { AuthController } from './auth/auth/auth.controller';
