+++ src/favorites/favorites.controller.ts
@@ -1,11 +1,19 @@
-import { Controller, Delete, Get, HttpCode, Param, Post, UseGuards } from '@nestjs/common';
+import {
+  Controller,
+  Delete,
+  Get,
+  HttpCode,
+  Param,
+  Post,
+  UseGuards,
+} from '@nestjs/common';
 import { FavoritesService } from './favorites.service';
 import { AuthGuard } from 'src/auth/auth/auth.quard';
 
 @Controller('favs')
 @UseGuards(AuthGuard)
 export class FavoritesController {
-  constructor(private favoritesService: FavoritesService) { }
+  constructor(private favoritesService: FavoritesService) {}
 
   @Get()
   findAll() {
