+++ src/logging/logger.middleware.ts
@@ -4,7 +4,7 @@ import { LoggingService } from './logging.service';
 
 @Injectable()
 export class LoggerMiddleware implements NestMiddleware {
-  constructor(private logger: LoggingService) { }
+  constructor(private logger: LoggingService) {}
 
   async use(req: Request, res: Response, next: NextFunction) {
     console.log('Request...');
